logout

+++ frontend/src/app/components/home/services/user/user.service.ts
@@ -8,10 +8,9 @@ import { toObservable } from '@angular/core/rxjs-interop';
 export class UserService {
   private isAuthenticated = signal<boolean>(false);
   private loggedInUserInfo = signal<LoggedInUser>({} as LoggedInUser);
-  private autoLogoutTimer: any;
 
   constructor(private http: HttpClient) {
-    //   this.loadToken();
+    this.loadToken();
   }
 
   get isUserAuthenticated(): boolean {
@@ -35,7 +34,6 @@ export class UserService {
   }
 
   activateToken(token: LoginToken): void {
-    //token.expiresInSeconds = 10;
     localStorage.setItem('token', token.token);
     localStorage.setItem(
       'expiry',
@@ -48,42 +46,37 @@ export class UserService {
 
     this.isAuthenticated.set(true);
     this.loggedInUserInfo.set(token.user);
-    //  this.setAutoLogoutTimer(token.expiresInSeconds * 1000);
   }
 
   logout(): void {
-    localStorage.clear();
-    this.isAuthenticated.set(false);
-    this.loggedInUserInfo.set({} as LoggedInUser);
-    clearTimeout(this.autoLogoutTimer);
+    if (typeof window !== 'undefined' && window.localStorage) {
+      localStorage.clear();
+      this.isAuthenticated.set(false);
+      this.loggedInUserInfo.set({} as LoggedInUser);
+    }
   }
 
-  private setAutoLogoutTimer(duration: number): void {
-    this.autoLogoutTimer = setTimeout(() => {
-      this.logout();
-    }, duration);
+  loadToken(): void {
+    if (typeof window !== 'undefined' && window.localStorage) {
+      const token = localStorage.getItem('token');
+      const expiry = localStorage.getItem('expiry');
+
+      if (!token || !expiry) return;
+
+      const expiresIn = new Date(expiry).getTime() - Date.now();
+      if (expiresIn > 0) {
+        const user: LoggedInUser = {
+          firstName: localStorage.getItem('firstName') || '',
+          lastName: localStorage.getItem('lastName') || '',
+          address: localStorage.getItem('address') || '',
+          city: localStorage.getItem('city') || '',
+        };
+
+        this.isAuthenticated.set(true);
+        this.loggedInUserInfo.set(user);
+      } else {
+        this.logout();
+      }
+    }
   }
-
-  // loadToken(): void {
-  //   const token = localStorage.getItem('token');
-  //   const expiry = localStorage.getItem('expiry');
-
-  //   if (!token || !expiry) return;
-
-  //   const expiresIn = new Date(expiry).getTime() - Date.now();
-  //   if (expiresIn > 0) {
-  //     const user: LoggedInUser = {
-  //       firstName: localStorage.getItem('firstName') || '',
-  //       lastName: localStorage.getItem('lastName') || '',
-  //       address: localStorage.getItem('address') || '',
-  //       city: localStorage.getItem('city') || '',
-  //     };
-
-  //     this.isAuthenticated.set(true);
-  //     this.loggedInUserInfo.set(user);
-  //     this.setAutoLogoutTimer(expiresIn);
-  //   } else {
-  //     this.logout();
-  //   }
-  // }
 }
